make change to the decorrelation part of the loss function
I split X and Xk into the discrete and continuous data. I’m not sure
whether the data is a torch object or a numpy object. Either way the
goal is to then calculate the pairwise correlations separately for the
discrete and continuous variables and we weight the discrete variables
more